requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-analytics-admin
google-analytics-data
httplib2
oauth2client
pandas
openpyxl
progress
protobuf
tabulate
urllib3
